Commit de avanco da aula 3 da NLW

+++ server/src/routes/memories.ts
@@ -4,9 +4,16 @@ import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
 
-    app.get('/memories', async () => {
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify()
+    })
+
+    app.get('/memories', async (request) => {
 
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub,
+            },
             orderBy: {
                 createdAt: 'asc',
             }
@@ -22,7 +29,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     });
 
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, reply) => {
 
         const paramsSchema = z.object({
             id: z.string().uuid(),
@@ -35,6 +42,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
             }
         })
 
+        if (!memory.isPublic && memory.userId != request.user.sub) {
+            return reply.status(401).send()
+        }
+
         return memory;
 
     });
@@ -46,27 +57,27 @@ export async function memoriesRoutes(app: FastifyInstance) {
             isPublic: z.coerce.boolean().default(false),
         })
 
-        const {content, coverUrl, isPublic} = bodySchema.parse(request.body);
+        const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
         const memory = await prisma.memory.create({
             data: {
                 content,
                 coverUrl,
                 isPublic,
-                userId: ''
+                userId: request.user.sub
             }
         })
 
         return memory;
     });
 
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, reply) => {
 
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
 
-        const {id} = paramsSchema.parse(request.params);
+        const { id } = paramsSchema.parse(request.params);
 
         const bodySchema = z.object({
             content: z.string(),
@@ -74,9 +85,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
             isPublic: z.coerce.boolean().default(false),
         })
 
-        const {content, coverUrl, isPublic} = bodySchema.parse(request.body)
+        const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findFirstOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if (memory.userId != request.user.sub){
+            return reply.status(401).send()
+        }
+
+        memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -91,12 +112,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     });
 
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, reply) => {
 
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
         const { id } = paramsSchema.parse(request.params);
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if (memory.userId != request.user.sub){
+            return reply.status(401).send()
+        }
 
         await prisma.memory.delete({
             where: {
@@ -104,7 +135,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
             }
         })
 
-       
+
 
     });
